Refactor feature-prompt to use feature-utils
